requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
black==19.10b0
Click==7.0
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
Faker==4.0.1
pathspec==0.7.0
python-dateutil==2.8.1
pytz==2019.3
regex==2020.2.20
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.0
typed-ast==1.4.1
